+++ Makefile.am
@@ -288,6 +288,7 @@ noinst_HEADERS = $(wsd_headers) $(shared_headers) $(kit_headers) \
                  test/countloolkits.hpp \
                  test/lokassert.hpp \
                  test/test.hpp \
+                 test/testlog.hpp \
                  test/helpers.hpp
 
 dist-hook:
